requirements.txt
Flask==3.1.0
Werkzeug==3.1.3
Jinja2==3.1.5

numpy==1.20.3
pandas==1.3.5
scipy==1.10.1
joblib==1.2.0
scikit-learn==1.3.0
xgboost==2.1.3

torch==2.5.1
torchaudio==2.5.1
torchvision==0.16.1
openai-whisper==20240930
transformers==4.47.1

pyttsx3==2.98
SpeechRecognition==3.13.0
PyAudio==0.2.14

requests==2.28.2

plotly==5.24.1
matplotlib==3.4.3
seaborn==0.13.2

python-dotenv==1.0.0
aiohttp==3.11.10